add `Value::Number` and a macro to parse
