+++ esp-hal/Cargo.toml
@@ -74,7 +74,7 @@ basic-toml   = "0.1.9"
 cfg-if       = "1.0.0"
 esp-build    = { version = "0.1.0", path = "../esp-build" }
 esp-metadata = { version = "0.3.0", path = "../esp-metadata" }
-esp-config   = { version = "0.1.0", path = "../esp-config" }
+esp-config   = { version = "0.1.0", path = "../esp-config", features = ["std"] }
 serde        = { version = "1.0.209", features = ["derive"] }
 
 [features]
