+++ esp-wifi/Cargo.toml
@@ -52,10 +52,12 @@ atomic-waker = { version = "1.1.2", default-features = false, features = [
   "portable-atomic",
 ] }
 bt-hci = { version = "0.1.0", optional = true }
+esp-config = { version = "0.1.0", path = "../esp-config" }
 
 [build-dependencies]
 toml-cfg     = "0.2.0"
 esp-build    = { version = "0.1.0", path = "../esp-build" }
+esp-config   = { version = "0.1.0", path = "../esp-config", features = ["std"] }
 esp-metadata = { version = "0.3.0", path = "../esp-metadata" }
 
 [features]
